[ESIMD] Use new simd_mask<N> class instead of simd<unsigned short,N> (intel/llvm-test-suite#388)
* [ESIMD] Add simd_mask test, fix other tests to use simd_mask<N>.
* [ESIMD] Fix unsupported simd_view += id<1> operation in test.
* Fix simd mask usage in ESIMD tests, stride=1 for 1-element select.

Complementary test change to intel/llvm#4230

+++ SYCL/ESIMD/usm_gather_scatter_rgba.cpp
@@ -42,7 +42,7 @@ struct Kernel {
     simd<uint32_t, VL> byteOffsets(0, STRIDE * sizeof(T) * NUM_RGBA_CHANNELS);
     simd<T, VL *numChannels> v =
         gather_rgba<T, VL, CH_MASK>(bufIn + global_offset, byteOffsets);
-    v += i;
+    v += (int)i;
 
     simd<uint16_t, VL> pred = 1;
     pred[VL - MASKED_LANE_NUM_REV] = 0; // mask out the last lane
